Changes config from a function that returns an objectinto a more aptly named BaseObservableMapDefaults class
