+++ src/observable/map/JoinedMap.ts
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservableMap, BaseObservableMapConfig} from "./BaseObservableMap";
-import {config, Mapper, Updater, Comparator, Filter} from "./config";
+import {BaseObservableMap} from "./BaseObservableMap";
+import {BaseObservableMapDefaults, Mapper, Updater, Comparator, Filter} from "./BaseObservableMapDefaults";
 import {FilteredMap} from "./FilteredMap";
 import {MappedMap} from "./MappedMap";
 import {SortedMapList} from "../list/SortedMapList.js";
@@ -23,14 +23,13 @@ import {SubscriptionHandle} from "../BaseObservable";
 
 
 export class JoinedMap<K, V> extends BaseObservableMap<K, V> {
+    private _defaults = new BaseObservableMapDefaults<K, V>();
     protected _sources: BaseObservableMap<K, V>[];
-    private _config: BaseObservableMapConfig<K, V>;
     private _subscriptions?: SourceSubscriptionHandler<K, V>[];
 
     constructor(sources: BaseObservableMap<K, V>[]) {
         super();
         this._sources = sources;
-        this._config = config<K, V>();
     }
 
     onAdd(source: BaseObservableMap<K, V>, key: K, value: V): void {
@@ -136,19 +135,19 @@ export class JoinedMap<K, V> extends BaseObservableMap<K, V> {
     }
 
     join(...otherMaps: Array<typeof this>): JoinedMap<K, V> {
-        return this._config.join(this, ...otherMaps);
+        return this._defaults.join(this, ...otherMaps);
     }
 
     mapValues(mapper: Mapper<V>, updater?: Updater<V>): MappedMap<K, V> {
-        return this._config.mapValues(this, mapper, updater);
+        return this._defaults.mapValues(this, mapper, updater);
     }
 
     sortValues(comparator: Comparator<V>): SortedMapList {
-        return this._config.sortValues(this, comparator);
+        return this._defaults.sortValues(this, comparator);
     }
 
     filterValues(filter: Filter<K, V>): FilteredMap<K, V> {
-        return this._config.filterValues(this, filter);
+        return this._defaults.filterValues(this, filter);
     }
 
 }
