+++ src/observable/map/LogMap.ts
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservableMap, BaseObservableMapConfig} from "./BaseObservableMap";
-import {config, Mapper, Updater, Comparator, Filter} from "./config";
+import {BaseObservableMap} from "./BaseObservableMap";
+import {BaseObservableMapDefaults, Mapper, Updater, Comparator, Filter} from "./BaseObservableMapDefaults";
 import {FilteredMap} from "./FilteredMap";
 import {MappedMap} from "./MappedMap";
 import {JoinedMap} from "./JoinedMap";
@@ -25,17 +25,15 @@ import {ILogItem, LabelOrValues} from "../../logging/types";
 import {LogLevel} from "../../logging/LogFilter";
 
 export class LogMap<K, V> extends BaseObservableMap<K, V> {
+    private _defaults = new BaseObservableMapDefaults<K, V>();
     private _source: BaseObservableMap<K, V>;
     private _subscription?: SubscriptionHandle;
     private _log: ILogItem;
-    private _config: BaseObservableMapConfig<K, V>
-
 
     constructor(source: BaseObservableMap<K, V>, log: ILogItem) {
         super();
         this._source = source;
         this._log = log;
-        this._config = config<K, V>();
     }
 
     private log(labelOrValues: LabelOrValues, logLevel?: LogLevel): ILogItem {
@@ -88,19 +86,19 @@ export class LogMap<K, V> extends BaseObservableMap<K, V> {
     }
 
     join(...otherMaps: Array<typeof this>): JoinedMap<K, V> {
-        return this._config.join(this, ...otherMaps);
+        return this._defaults.join(this, ...otherMaps);
     }
 
     mapValues(mapper: Mapper<V>, updater?: Updater<V>): MappedMap<K, V> {
-        return this._config.mapValues(this, mapper, updater);
+        return this._defaults.mapValues(this, mapper, updater);
     }
 
     sortValues(comparator: Comparator<V>): SortedMapList {
-        return this._config.sortValues(this, comparator);
+        return this._defaults.sortValues(this, comparator);
     }
 
     filterValues(filter: Filter<K, V>): FilteredMap<K, V> {
-        return this._config.filterValues(this, filter);
+        return this._defaults.filterValues(this, filter);
     }
 
 }
